requirements.txt
# Core
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Integration
openai>=1.0.0
tenacity>=8.2.0

# Document Processing
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
pymupdf>=1.23.0

# Vision/OCR
easyocr>=1.7.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
